python-backend/requirements.txt
# SmartSpec Pro Python Backend - Requirements
# Phase 0: Technical Foundation

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.7.4
pydantic-settings>=2.3.0

# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# LLM Providers
openai>=1.50.0
anthropic==0.8.1
google-generativeai==0.3.2
groq==0.4.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis (for caching & state)
redis==5.0.1
hiredis==2.3.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==42.0.0

# HTTP & Async
httpx>=0.23.0,<1.0
aiohttp==3.9.1
aiofiles==23.2.1

# Utilities
pyyaml==6.0.1
toml==0.10.2
python-dateutil==2.8.2
pytz==2023.3
psutil==5.9.8

# Logging & Monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
# pytest-httpx - removed due to dependency conflict

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3

# Background Tasks (for Email)
dramatiq[redis]==1.15.0

itsdangerous==2.1.2

PyJWT==2.8.0

aiosqlite==0.19.0

stripe==8.5.0

aiosmtplib==2.0.1

argon2-cffi==23.1.0

chromadb>=0.5.0
